+++ src/neptune/common/backends/utils.py
@@ -34,6 +34,7 @@ from bravado.exception import (
     HTTPTooManyRequests,
     HTTPUnauthorized,
 )
+from bravado_core.util import RecursiveCallException
 from requests.exceptions import ChunkedEncodingError
 from urllib3.exceptions import NewConnectionError
 
@@ -44,10 +45,8 @@ from neptune.common.exceptions import (
     NeptuneAuthTokenExpired,
     NeptuneConnectionLostException,
     NeptuneInvalidApiTokenException,
-    NeptuneSSLVerificationError,
     Unauthorized,
 )
-from neptune.common.utils import reset_internal_ssl_state
 
 _logger = logging.getLogger(__name__)
 
@@ -67,7 +66,6 @@ def get_retry_from_headers_or_default(headers, retry_count):
 
 def with_api_exceptions_handler(func):
     def wrapper(*args, **kwargs):
-        ssl_error_occurred = False
         last_exception = None
         start_time = time.monotonic()
         for retry in itertools.count(0):
@@ -80,22 +78,6 @@ def with_api_exceptions_handler(func):
                 if "X-Neptune-Api-Token" in e.args[0]:
                     raise NeptuneInvalidApiTokenException()
                 raise
-            except requests.exceptions.SSLError as e:
-                """
-                OpenSSL's internal random number generator does not properly handle forked processes.
-                Applications must change the PRNG state of the parent process
-                if they use any SSL feature with os.fork().
-                Any successful call of RAND_add(), RAND_bytes() or RAND_pseudo_bytes() is sufficient.
-                https://docs.python.org/3/library/ssl.html#multi-processing
-
-                On Linux it looks like it does not help much but does not break anything either.
-                But single retry seems to solve the issue.
-                """
-                if not ssl_error_occurred:
-                    ssl_error_occurred = True
-                    reset_internal_ssl_state()
-                    continue
-                raise NeptuneSSLVerificationError() from e
             except (
                 BravadoConnectionError,
                 BravadoTimeoutError,
@@ -108,6 +90,8 @@ def with_api_exceptions_handler(func):
                 HTTPInternalServerError,
                 NewConnectionError,
                 ChunkedEncodingError,
+                requests.exceptions.SSLError,
+                RecursiveCallException,
             ) as e:
                 time.sleep(min(2 ** min(MAX_RETRY_MULTIPLIER, retry), MAX_RETRY_TIME))
                 last_exception = e