+++ src/neptune/internal/operation_processors/async_operation_processor.py
@@ -26,38 +26,30 @@ from time import (
 )
 from typing import (
     Callable,
-    ClassVar,
     List,
     Optional,
-    Tuple,
 )
 
 from neptune.constants import ASYNC_DIRECTORY
 from neptune.envs import NEPTUNE_SYNC_AFTER_STOP_TIMEOUT
 from neptune.exceptions import NeptuneSynchronizationAlreadyStoppedException
 from neptune.internal.backends.neptune_backend import NeptuneBackend
-from neptune.internal.backends.operations_preprocessor import OperationsPreprocessor
 from neptune.internal.container_type import ContainerType
+from neptune.internal.disk_queue import DiskQueue
 from neptune.internal.id_formats import UniqueId
 from neptune.internal.init.parameters import (
     ASYNC_LAG_THRESHOLD,
     ASYNC_NO_PROGRESS_THRESHOLD,
     DEFAULT_STOP_TIMEOUT,
 )
-from neptune.internal.operation import (
-    CopyAttribute,
-    Operation,
-)
+from neptune.internal.operation import Operation
 from neptune.internal.operation_processors.operation_processor import OperationProcessor
 from neptune.internal.operation_processors.operation_storage import (
     OperationStorage,
     get_container_dir,
 )
-from neptune.internal.queue.disk_queue import (
-    DiskQueue,
-    QueueElement,
-)
 from neptune.internal.threading.daemon import Daemon
+from neptune.internal.utils.disk_full import ensure_disk_not_full
 from neptune.internal.utils.logger import logger
 
 _logger = logging.getLogger(__name__)
@@ -101,7 +93,7 @@ class AsyncOperationProcessor(OperationProcessor):
         self._consumed_version = 0
         self._consumer = self.ConsumerThread(self, sleep_time, batch_size)
         self._lock = lock
-        self._last_ack = None
+        self._last_ack: Optional[float] = None
         self._lag_exceeded = False
         self._should_call_no_progress_callback = False
 
@@ -114,6 +106,7 @@ class AsyncOperationProcessor(OperationProcessor):
         process_path = f"exec-{now.timestamp()}-{now.strftime('%Y-%m-%d_%H.%M.%S.%f')}-{os.getpid()}"
         return get_container_dir(ASYNC_DIRECTORY, container_id, container_type, process_path)
 
+    @ensure_disk_not_full
     def enqueue_operation(self, op: Operation, *, wait: bool) -> None:
         self._last_version = self._queue.put(op)
 
@@ -264,10 +257,6 @@ class AsyncOperationProcessor(OperationProcessor):
         self._queue.close()
 
     class ConsumerThread(Daemon):
-        MAX_OPERATIONS_IN_BATCH: ClassVar[int] = 1000
-        MAX_APPENDS_IN_BATCH: ClassVar[int] = 100000
-        MAX_BATCH_SIZE_BYTES: ClassVar[int] = 100 * 1024 * 1024
-
         def __init__(
             self,
             processor: "AsyncOperationProcessor",
@@ -279,7 +268,6 @@ class AsyncOperationProcessor(OperationProcessor):
             self._batch_size = batch_size
             self._last_flush = 0
             self._no_progress_exceeded = False
-            self._last_disk_record: Optional[QueueElement[Operation]] = None
 
         def run(self):
             try:
@@ -296,46 +284,17 @@ class AsyncOperationProcessor(OperationProcessor):
                 self._processor._queue.flush()
 
             while True:
-                batch = self.collect_batch()
+                batch = self._processor._queue.get_batch(self._batch_size)
                 if not batch:
                     return
-                operations, version = batch
-                self.process_batch(operations, version)
-
-        def collect_batch(self) -> Optional[Tuple[List[Operation], int]]:
-            preprocessor = OperationsPreprocessor()
-            version: Optional[int] = None
-            total_bytes = 0
-            copy_ops: List[CopyAttribute] = []
-            while (
-                preprocessor.final_ops_count < self.MAX_OPERATIONS_IN_BATCH
-                and preprocessor.final_append_count < self.MAX_APPENDS_IN_BATCH
-                and total_bytes < self.MAX_BATCH_SIZE_BYTES
-            ):
-                record: Optional[QueueElement[Operation]] = self._last_disk_record or self._processor._queue.get()
-                self._last_disk_record = None
-                if not record:
-                    break
-                if isinstance(record.obj, CopyAttribute):
-                    # CopyAttribute can be only at the start of a batch.
-                    if copy_ops or preprocessor.final_ops_count:
-                        self._last_disk_record = record
-                        break
-                    else:
-                        version = record.ver
-                        copy_ops.append(record.obj)
-                        total_bytes += record.size
-                elif preprocessor.process(record.obj):
-                    version = record.ver
-                    total_bytes += record.size
-                else:
-                    self._last_disk_record = record
-                    break
-            return (copy_ops + preprocessor.get_operations().all_operations(), version) if version is not None else None
+                self.process_batch([element.obj for element in batch], batch[-1].ver)
 
         def _check_no_progress(self):
             if not self._no_progress_exceeded:
-                if monotonic() - self._processor._last_ack > self._processor._async_no_progress_threshold:
+                if (
+                    self._processor._last_ack
+                    and monotonic() - self._processor._last_ack > self._processor._async_no_progress_threshold
+                ):
                     self._no_progress_exceeded = True
                     self._processor._should_call_no_progress_callback = True