+++ src/neptune/internal/utils/json_file_splitter.py
@@ -15,20 +15,13 @@
 #
 __all__ = ["JsonFileSplitter"]
 
+import json
 from collections import deque
 from io import StringIO
-from json import (
-    JSONDecodeError,
-    JSONDecoder,
-)
-from types import TracebackType
+from json import JSONDecodeError
 from typing import (
-    IO,
-    Any,
-    Deque,
     Optional,
     Tuple,
-    Type,
 )
 
 
@@ -37,11 +30,11 @@ class JsonFileSplitter:
     MAX_PART_READ = 8 * 1024
 
     def __init__(self, file_path: str):
-        self._file: IO = open(file_path, "r")
-        self._decoder: JSONDecoder = JSONDecoder(strict=False)
-        self._part_buffer: StringIO = StringIO()
-        self._parsed_queue: Deque[Tuple[Any, int]] = deque()
-        self._start_pos: int = 0
+        self._file = open(file_path, "r")
+        self._decoder = json.JSONDecoder(strict=False)
+        self._part_buffer = StringIO()
+        self._parsed_queue = deque()
+        self._start_pos = 0
 
     def close(self) -> None:
         self._file.close()
@@ -58,7 +51,7 @@ class JsonFileSplitter:
             return self._parsed_queue.popleft()
         return None, 0
 
-    def _read_data(self) -> None:
+    def _read_data(self):
         if self._part_buffer.tell() < self.MAX_PART_READ:
             data = self._file.read(self.BUFFER_SIZE)
             if not data:
@@ -75,7 +68,7 @@ class JsonFileSplitter:
             data = self._reset_part_buffer()
             self._decode(data)
 
-    def _decode(self, data: str) -> None:
+    def _decode(self, data: str):
         start = self._json_start(data)
         while start is not None:
             try:
@@ -101,11 +94,3 @@ class JsonFileSplitter:
         self._part_buffer.close()
         self._part_buffer = StringIO()
         return data
-
-    def __enter__(self) -> "JsonFileSplitter":
-        return self
-
-    def __exit__(
-        self, exc_type: Type[BaseException], exc_value: BaseException, exc_traceback: Optional[TracebackType]
-    ) -> None:
-        self.close()